Fix EnrollmentAnalyticsManagerTest
