+++ dhis-2/dhis-services/dhis-service-analytics/src/test/java/org/hisp/dhis/analytics/event/data/EnrollmentAnalyticsManagerCteTest.java
@@ -32,16 +32,22 @@ package org.hisp.dhis.analytics.event.data;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hisp.dhis.analytics.DataType.NUMERIC;
 import static org.hisp.dhis.analytics.QueryKey.NV;
 import static org.hisp.dhis.analytics.table.EventAnalyticsColumnName.OCCURRED_DATE_COLUMN_NAME;
 import static org.hisp.dhis.analytics.table.EventAnalyticsColumnName.OU_COLUMN_NAME;
 import static org.hisp.dhis.common.DimensionConstants.OPTION_SEP;
+import static org.hisp.dhis.common.QueryOperator.EQ;
 import static org.hisp.dhis.common.QueryOperator.IN;
+import static org.hisp.dhis.common.QueryOperator.NEQ;
 import static org.hisp.dhis.common.RequestTypeAware.EndpointAction.AGGREGATE;
 import static org.hisp.dhis.external.conf.ConfigurationKey.ANALYTICS_DATABASE;
 import static org.hisp.dhis.program.EnrollmentStatus.ACTIVE;
 import static org.hisp.dhis.program.EnrollmentStatus.COMPLETED;
 import static org.hisp.dhis.test.TestBase.createPeriodDimensions;
+import static org.hisp.dhis.test.TestBase.createProgram;
+import static org.hisp.dhis.test.TestBase.createProgramIndicator;
+import static org.hisp.dhis.test.TestBase.getDate;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -81,10 +87,16 @@ import org.hisp.dhis.db.sql.DorisAnalyticsSqlBuilder;
 import org.hisp.dhis.db.sql.PostgreSqlAnalyticsSqlBuilder;
 import org.hisp.dhis.external.conf.DefaultDhisConfigurationProvider;
 import org.hisp.dhis.period.PeriodDimension;
+import org.hisp.dhis.program.Program;
+import org.hisp.dhis.program.ProgramIndicator;
 import org.hisp.dhis.program.ProgramIndicatorService;
+import org.hisp.dhis.relationship.RelationshipConstraint;
+import org.hisp.dhis.relationship.RelationshipEntity;
+import org.hisp.dhis.relationship.RelationshipType;
 import org.hisp.dhis.setting.SystemSettings;
 import org.hisp.dhis.setting.SystemSettingsService;
 import org.hisp.dhis.system.grid.ListGrid;
+import org.hisp.dhis.test.random.BeanRandomizer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -128,6 +140,8 @@ class EnrollmentAnalyticsManagerCteTest extends EventAnalyticsTest {
   @Mock private PiDisagInfoInitializer piDisagInfoInitializer;
 
   @Mock private PiDisagQueryGenerator piDisagQueryGenerator;
+
+  private final BeanRandomizer rnd = BeanRandomizer.create();
 
   private QueryItemFilterBuilder filterBuilder;
 
@@ -708,6 +722,7 @@ class EnrollmentAnalyticsManagerCteTest extends EventAnalyticsTest {
     DefaultProgramIndicatorSubqueryBuilder programIndicatorSubqueryBuilder =
         new DefaultProgramIndicatorSubqueryBuilder(
             programIndicatorService, systemSettingsService, builder, dataElementService);
+    programIndicatorSubqueryBuilder.init();
     ColumnMapper columnMapper = new ColumnMapper(builder, systemSettingsService);
     filterBuilder = new QueryItemFilterBuilder(organisationUnitResolver, builder);
     EnrollmentTimeFieldSqlRenderer timeFieldRenderer = new EnrollmentTimeFieldSqlRenderer(builder);
@@ -764,6 +779,310 @@ class EnrollmentAnalyticsManagerCteTest extends EventAnalyticsTest {
         containsString("(completeddate >= '2022-09-01' and completeddate < '2023-09-01')"));
     assertThat(generatedSql, not(containsString("enrollmentdate >= '2022-09-01'")));
     assertThat(generatedSql, not(containsString("occurreddate >= '2022-09-01'")));
+  }
+
+  @Test
+  void verifyWithProgramAndStartEndDate() {
+    EventQueryParams params =
+        new EventQueryParams.Builder(createRequestParams())
+            .withStartDate(getDate(2017, 1, 1))
+            .withEndDate(getDate(2017, 12, 31))
+            .build();
+
+    subject.getEnrollments(params, new ListGrid(), 0);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = sql.getValue();
+    assertThat(generatedSql, containsString("enrollmentdate >= '2017-01-01'"));
+    assertThat(generatedSql, containsString("enrollmentdate < '2018-01-01'"));
+    assertThat(generatedSql, containsString("ax.\"uidlevel1\" = 'ouabcdefghA'"));
+  }
+
+  @Test
+  void verifyWithLastUpdatedTimeField() {
+    EventQueryParams params =
+        new EventQueryParams.Builder(createRequestParams())
+            .withStartDate(getDate(2017, 1, 1))
+            .withEndDate(getDate(2017, 12, 31))
+            .withTimeField(TimeField.LAST_UPDATED.name())
+            .build();
+
+    subject.getEnrollments(params, new ListGrid(), 10000);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = sql.getValue();
+    assertThat(generatedSql, containsString("lastupdated >= '2017-01-01'"));
+    assertThat(generatedSql, containsString("lastupdated < '2018-01-01'"));
+  }
+
+  @Test
+  void verifyWithProgramStageAndNumericDataElement() {
+    verifyWithProgramStageAndDataElement(ValueType.NUMBER);
+  }
+
+  @Test
+  void verifyWithProgramStageAndTextDataElement() {
+    verifyWithProgramStageAndDataElement(ValueType.TEXT);
+  }
+
+  private void verifyWithProgramStageAndDataElement(ValueType valueType) {
+    EventQueryParams params = createRequestParams(this.programStage, valueType);
+
+    subject.getEnrollments(params, new ListGrid(), 100);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = noEof(sql.getValue());
+    String stageUid = programStage.getUid();
+    String deUid = dataElementA.getUid();
+
+    // CTE for the stage data element
+    assertThat(
+        generatedSql,
+        containsString(
+            noEof(
+                """
+                %s_%s_0 as ( select enrollment, "%s" as value,
+                row_number() over ( partition by enrollment order by occurreddate desc, created desc ) as rn
+                from analytics_event_%s where eventstatus != 'SCHEDULE' and ps = '%s' )
+                """
+                    .formatted(stageUid, deUid, deUid, programA.getUid(), stageUid))));
+
+    // Value projected as stage.de alias
+    assertThat(generatedSql, containsString("as \"" + stageUid + "." + deUid + "\""));
+
+    // Stage filter in WHERE
+    assertThat(generatedSql, containsString("and ps = '" + stageUid + "'"));
+  }
+
+  @Test
+  void verifyWithRepeatableProgramStageAndNumericDataElement() {
+    verifyWithRepeatableProgramStageAndDataElement(ValueType.NUMBER);
+  }
+
+  @Test
+  void verifyWithRepeatableProgramStageAndTextDataElement() {
+    verifyWithRepeatableProgramStageAndDataElement(ValueType.TEXT);
+  }
+
+  private void verifyWithRepeatableProgramStageAndDataElement(ValueType valueType) {
+    EventQueryParams params = createRequestParams(repeatableProgramStage, valueType);
+
+    subject.getEnrollments(params, new ListGrid(), 100);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = noEof(sql.getValue());
+    String stageUid = repeatableProgramStage.getUid();
+    String deUid = dataElementA.getUid();
+
+    // CTE includes eventstatus for repeatable stage
+    assertThat(
+        generatedSql,
+        containsString(
+            noEof(
+                """
+                %s_%s_0 as ( select enrollment, "%s" as value, eventstatus,
+                row_number() over ( partition by enrollment order by occurreddate desc, created desc ) as rn
+                from analytics_event_%s where eventstatus != 'SCHEDULE' and ps = '%s' )
+                """
+                    .formatted(stageUid, deUid, deUid, programA.getUid(), stageUid))));
+
+    // Existence CTE
+    assertThat(
+        generatedSql,
+        containsString(
+            noEof(
+                """
+                select distinct enrollment from analytics_event_%s
+                where eventstatus != 'SCHEDULE' and ps = '%s'
+                """
+                    .formatted(programA.getUid(), stageUid))));
+
+    // SELECT projections for repeatable stage
+    assertThat(generatedSql, containsString("\"" + stageUid + "[-1]." + deUid + "\""));
+    assertThat(generatedSql, containsString("\"" + stageUid + "[-1]." + deUid + ".exists\""));
+    assertThat(generatedSql, containsString("\"" + stageUid + "[-1]." + deUid + ".status\""));
+  }
+
+  @Test
+  void verifyWithProgramStageAndTextualDataElementAndFilter() {
+    verifyWithProgramStageAndDataElementAndFilter(ValueType.TEXT);
+  }
+
+  @Test
+  void verifyWithProgramStageAndNumericDataElementAndFilter() {
+    verifyWithProgramStageAndDataElementAndFilter(ValueType.NUMBER);
+  }
+
+  private void verifyWithProgramStageAndDataElementAndFilter(ValueType valueType) {
+    EventQueryParams params = createRequestParamsWithFilter(programStage, valueType);
+
+    subject.getEnrollments(params, new ListGrid(), 10000);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = noEof(sql.getValue());
+    String stageUid = programStage.getUid();
+    String deUid = dataElementA.getUid();
+
+    // Filter pushed into CTE
+    assertThat(
+        generatedSql,
+        containsString(
+            noEof(
+                """
+                %s_%s_0 as ( select enrollment, "%s" as value,
+                row_number() over ( partition by enrollment order by occurreddate desc, created desc ) as rn
+                from analytics_event_%s where eventstatus != 'SCHEDULE' and ps = '%s' and "%s" > '10' )
+                """
+                    .formatted(stageUid, deUid, deUid, programA.getUid(), stageUid, deUid))));
+
+    // Inner join (not left join) because filter exists
+    assertThat(generatedSql, containsString("inner join " + stageUid + "_" + deUid + "_0"));
+  }
+
+  @Test
+  void verifyGetEventsWithProgramStatusParam() {
+    mockEmptyRowSet();
+    EventQueryParams params = createRequestParamsWithStatusesForEnrollmentQuery();
+
+    subject.getEnrollments(params, new ListGrid(), 10000);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    assertThat(sql.getValue(), containsString("enrollmentstatus in ('ACTIVE','COMPLETED')"));
+  }
+
+  @Test
+  void verifyGetEnrollmentsWithMissingValueEqFilter() {
+    String stageUid = programStage.getUid();
+    String deUid = dataElementA.getUid();
+
+    // CTE without filter (NV stays in WHERE)
+    String cte =
+        noEof(
+            """
+            %s_%s_0 as ( select enrollment, "%s" as value,
+            row_number() over ( partition by enrollment order by occurreddate desc, created desc ) as rn
+            from analytics_event_%s where eventstatus != 'SCHEDULE' and ps = '%s' )
+            """
+                .formatted(stageUid, deUid, deUid, programA.getUid(), stageUid));
+
+    testIt(
+        EQ,
+        NV,
+        List.of(
+            capturedSql -> assertThat(capturedSql, containsString(cte)),
+            capturedSql -> assertThat(capturedSql, containsString("value is NULL"))));
+  }
+
+  @Test
+  void verifyGetEnrollmentsWithMissingValueNeqFilter() {
+    String stageUid = programStage.getUid();
+    String deUid = dataElementA.getUid();
+
+    String cte =
+        noEof(
+            """
+            %s_%s_0 as ( select enrollment, "%s" as value,
+            row_number() over ( partition by enrollment order by occurreddate desc, created desc ) as rn
+            from analytics_event_%s where eventstatus != 'SCHEDULE' and ps = '%s' )
+            """
+                .formatted(stageUid, deUid, deUid, programA.getUid(), stageUid));
+
+    testIt(
+        NEQ,
+        NV,
+        List.of(
+            capturedSql -> assertThat(capturedSql, containsString(cte)),
+            capturedSql -> assertThat(capturedSql, containsString("value is not NULL"))));
+  }
+
+  @Test
+  void verifyWithProgramIndicatorAndRelationshipTypeBothSidesTrackedEntity() {
+    ProgramIndicator programIndicatorA = createProgramIndicator('A', programA, "", "");
+
+    RelationshipType relationshipTypeA = createRelationshipType();
+
+    EventQueryParams.Builder params =
+        new EventQueryParams.Builder(createRequestParams(programIndicatorA, relationshipTypeA))
+            .withStartDate(getDate(2015, 1, 1))
+            .withEndDate(getDate(2017, 4, 8));
+
+    when(programIndicatorService.getAnalyticsSql(
+            "", NUMERIC, programIndicatorA, getDate(2000, 1, 1), getDate(2017, 4, 8), "subax"))
+        .thenReturn("distinct event");
+
+    subject.getEnrollments(params.build(), new ListGrid(), 100);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = sql.getValue();
+    assertThat(generatedSql, containsString("value as " + programIndicatorA.getUid()));
+    assertThat(generatedSql, containsString("enrollmentdate >= '2015-01-01'"));
+    assertThat(generatedSql, containsString("enrollmentdate < '2017-04-09'"));
+  }
+
+  @Test
+  void verifyWithProgramIndicatorAndRelationshipTypeDifferentConstraint() {
+    ProgramIndicator programIndicatorA = createProgramIndicator('A', programA, "", "");
+
+    RelationshipType relationshipTypeA =
+        createRelationshipType(RelationshipEntity.PROGRAM_INSTANCE);
+
+    EventQueryParams.Builder params =
+        new EventQueryParams.Builder(createRequestParams(programIndicatorA, relationshipTypeA))
+            .withStartDate(getDate(2015, 1, 1))
+            .withEndDate(getDate(2017, 4, 8));
+
+    when(programIndicatorService.getAnalyticsSql(
+            "", NUMERIC, programIndicatorA, getDate(2000, 1, 1), getDate(2017, 4, 8), "subax"))
+        .thenReturn("distinct event");
+
+    subject.getEnrollments(params.build(), new ListGrid(), 100);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = sql.getValue();
+    assertThat(generatedSql, containsString("value as " + programIndicatorA.getUid()));
+    assertThat(generatedSql, containsString("enrollmentdate >= '2015-01-01'"));
+  }
+
+  @Test
+  void verifyWithProgramIndicatorAndRelationshipTypeBothSidesTrackedEntity2() {
+    Program programB = createProgram('B');
+    ProgramIndicator programIndicatorA = createProgramIndicator('A', programB, "", "");
+
+    RelationshipType relationshipTypeA = createRelationshipType();
+
+    EventQueryParams.Builder params =
+        new EventQueryParams.Builder(createRequestParams(programIndicatorA, relationshipTypeA))
+            .withStartDate(getDate(2015, 1, 1))
+            .withEndDate(getDate(2017, 4, 8));
+
+    when(programIndicatorService.getAnalyticsSql(
+            "", NUMERIC, programIndicatorA, getDate(2000, 1, 1), getDate(2017, 4, 8), "subax"))
+        .thenReturn("distinct event");
+
+    subject.getEnrollments(params.build(), new ListGrid(), 100);
+    verify(jdbcTemplate).queryForRowSet(sql.capture());
+
+    String generatedSql = sql.getValue();
+    assertThat(generatedSql, containsString("value as " + programIndicatorA.getUid()));
+    assertThat(generatedSql, containsString("enrollmentdate >= '2015-01-01'"));
+  }
+
+  private RelationshipType createRelationshipType(RelationshipEntity toConstraint) {
+    RelationshipType relationshipTypeA = rnd.nextObject(RelationshipType.class);
+
+    RelationshipConstraint from = new RelationshipConstraint();
+    from.setRelationshipEntity(RelationshipEntity.TRACKED_ENTITY_INSTANCE);
+
+    RelationshipConstraint to = new RelationshipConstraint();
+    to.setRelationshipEntity(toConstraint);
+
+    relationshipTypeA.setFromConstraint(from);
+    relationshipTypeA.setToConstraint(to);
+    return relationshipTypeA;
+  }
+
+  private RelationshipType createRelationshipType() {
+    return createRelationshipType(RelationshipEntity.TRACKED_ENTITY_INSTANCE);
   }
 
   private EventQueryParams createAggregateEnrollmentWithStageDateParams() {
